data type fixed

+++ src/lib.rs
@@ -1,21 +1,21 @@
 use std::ops::Add;
 
-pub fn add(num1: u32, num2: u32) -> u32 {
+pub fn add(num1: i64, num2: i64) -> i64 {
     num1 + num2
 }
 
 
-pub fn subtract(num1: u32, num2: u32) -> u32 {
+pub fn subtract(num1: i64, num2: i64) -> i64 {
     num1 - num2
 }
 
 
-pub fn multiply(num1: u32, num2: u32) -> u32 {
+pub fn multiply(num1: i64, num2: i64) -> i64 {
     num1 * num2
 }
 
 
-pub fn divide(num1: u32, num2: u32) -> u32 {
+pub fn divide(num1: i64, num2: i64) -> i64 {
     if (num2 == 0) {
         println!("can not divide with zero");
         0
@@ -25,6 +25,6 @@ pub fn divide(num1: u32, num2: u32) -> u32 {
 }
 
 
-pub fn power(num1: u32, num2: u32) -> u32 {
+pub fn power(num1: i64, num2: u32) -> i64 {
     num1.pow(num2)
 }
